% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridstackr.R
\name{addWidget}
\alias{addWidget}
\title{Adds a new widget to the gridstack with element id}
\usage{
addWidget(gridstackrProxy)
}
\arguments{
\item{gridstackrProxy}{}
}
\value{
gridstackrProxy
}
\description{
Adds a new widget to the gridstack with element id
}
