% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{create_design}
\alias{create_design}
\title{Create a patchwork design}
\usage{
create_design(grid_layout)
}
\arguments{
\item{grid_layout}{Layout of the GridStack retrieved via \verb{input$<outputId>_layout}.}
}
\value{
An expression that can be evaluated to create a design usable in {patchwork}.
}
\description{
Create a patchwork design
}
\examples{

if (require("bslib", quietly=TRUE) &
    require("ggplot2", quietly=TRUE) &
    require("patchwork", quietly=TRUE)) {

  library(shiny)
  library(bslib)
  library(ggplot2)
  library(gridstackr)
  library(patchwork)

  ui <- page_fluid(
    tags$h2("Create {patchwork} design"),
    fluidRow(
      column(
        width = 6,
        tags$b("Create your design:"),
        gridstackOutput("grid")
      ),
      column(
        width = 6,
        tags$b("Result:"),
        plotOutput("design")
      )
    )
  )

  server <- function(input, output, session) {

    output$grid <- renderGridstack({
      gridstack(
        resize_handles = "se,e,s",
        gs_item(
          plotOutput("plot1", height = "100\%")
        ),
        gs_item(
          plotOutput("plot2", height = "100\%")
        ),
        gs_item(
          plotOutput("plot3", height = "100\%")
        ),
        gs_item(
          plotOutput("plot4", height = "100\%")
        )
      )
    })

    output$plot1 <- renderPlot({
      ggplot(mtcars) + geom_point(aes(mpg, disp))
    })
    outputOptions(output, "plot1", suspendWhenHidden = TRUE)
    output$plot2 <- renderPlot({
      ggplot(mtcars) + geom_boxplot(aes(gear, disp, group = gear))
    })
    output$plot3 <- renderPlot({
      ggplot(mtcars) + geom_smooth(aes(disp, qsec))
    })
    output$plot4 <- renderPlot({
      ggplot(mtcars) + geom_bar(aes(carb))
    })

    output$design <- renderPlot({
      req(input$grid_layout)
      mydesign <- create_design(input$grid_layout)
      plot(eval(mydesign))
    })

  }

  if (interactive())
    shinyApp(ui, server)

}
}
