% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridstackr.R
\name{gridstackrProxy}
\alias{gridstackrProxy}
\title{Create a gridstackr proxy object}
\usage{
gridstackrProxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Name of the gridstackr widget}

\item{session}{Valid session object}
}
\value{
grid proxy object
}
\description{
Create a gridstackr proxy object
}
