% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridstack.R
\name{gs_item}
\alias{gs_item}
\title{Gridstack Item}
\usage{
gs_item(..., x = 0, y = 0, w = 4, h = 4)
}
\arguments{
\item{...}{Elements to include within the grid item.}

\item{x}{x position to put the grid item.}

\item{y}{y position to put the grid item.}

\item{w}{weight of the grid item.}

\item{h}{height  of the grid item.}
}
\description{
Gridstack Item
}

