% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridstackr.R
\name{initGridstackr}
\alias{initGridstackr}
\title{Initialization for gridstackr object}
\usage{
initGridstackr(ui, server, env = parent.frame(), .ns = NS("init"))
}
\arguments{
\item{ui}{UI code}

\item{server}{Server code}

\item{env}{Environment}

\item{.ns}{Namespace (from call to shiny::NS)}
}
\description{
Initialization for gridstackr object
}
