% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{proxy-methods}
\alias{proxy-methods}
\alias{gs_proxy_compact}
\alias{gs_proxy_enable}
\alias{gs_proxy_disable}
\alias{gs_proxy_enable_move}
\alias{gs_proxy_enable_resize}
\title{Proxy methods}
\usage{
gs_proxy_compact(proxy, layout = c("list", "compact"))

gs_proxy_enable(proxy)

gs_proxy_disable(proxy)

gs_proxy_enable_move(proxy, doEnable = TRUE)

gs_proxy_enable_resize(proxy, doEnable = TRUE)
}
\arguments{
\item{proxy}{Result of \code{\link[=gridstack_proxy]{gridstack_proxy()}} or a character with the ID of the grid.}

\item{layout}{re-layout grid items to reclaim any empty space. Options are:
\itemize{
\item 'list' keep the widget left->right order the same, even if that means leaving an empty slot if things don't fit
\item 'compact' might re-order items to fill any empty space.
}}

\item{doEnable}{Enables/disables widget moving or resizing.}
}
\value{
A \code{\link[=gridstack_proxy]{gridstack_proxy()}} object.
}
\description{
Call those functions from the server of an application to interact with a GridStack.
}
\examples{

library(shiny)
library(bslib)
library(gridstackr)

ui <- page_fluid(
  tags$h2("GridStack Input example"),
  gridstackOutput("mygrid"),
  actionButton("compact", "Compact"),
  actionButton("enable", "Enable"),
  actionButton("disable", "Disable"),
  actionButton("enableMove", "Enable Move"),
  actionButton("disableMove", "Disable Move"),
  actionButton("enableResize", "Enable Resize"),
  actionButton("disableResize", "Disable Resize")
)

server <- function(input, output, session) {

  output$mygrid <- renderGridstack({
    gridstack(
      minRow = 2,
      margin = "0.2rem",
      gs_item("1", id = "item_1", w = 2, h = 2, class_content = "gs-item-example"),
      gs_item("2", id = "item_2", w = 4, class_content = "gs-item-example"),
      gs_item("3", id = "item_3", w = 1, x = 4, y = 2, class_content = "gs-item-example"),
      gs_item("4", id = "item_4", h = 2, w = 2, class_content = "gs-item-example"),
      gs_item("5", id = "item_5", w = 2, h = 1, x = 0, y = 3, class_content = "gs-item-example"),
      gs_item("6", id = "item_6", w = 3, x = 7, y = 3, class_content = "gs-item-example")
    )
  })

  observeEvent(input$compact, gs_proxy_compact("mygrid"))
  observeEvent(input$enable, gs_proxy_enable("mygrid"))
  observeEvent(input$disable, gs_proxy_disable("mygrid"))
  observeEvent(input$enableMove, gs_proxy_enable_move("mygrid", TRUE))
  observeEvent(input$disableMove, gs_proxy_enable_move("mygrid", FALSE))
  observeEvent(input$enableResize, gs_proxy_enable_resize("mygrid", TRUE))
  observeEvent(input$disableResize, gs_proxy_enable_resize("mygrid", FALSE))

}

if (interactive())
  shinyApp(ui, server)
}
