% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridstackr.R
\name{removeWidget}
\alias{removeWidget}
\title{Removes a widget from the gridstack}
\usage{
removeWidget(gridstackrProxy, id = "", el = "")
}
\arguments{
\item{id}{Name of the gridstackr widget}

\item{el}{Element to be removed}
}
\value{
gridstackrProxy
}
\description{
Removes a widget from the gridstack
}
