% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridstack.R
\name{gridstack}
\alias{gridstack}
\title{Gridstack HTMLwidget}
\usage{
gridstack(..., options = list(), width = NULL, height = NULL, elementId = NULL)
}
\description{
Mobile-friendly modern Typescript library for dashboard layout and creation. Making a drag-and-drop, multi-column responsive dashboard has never been easier.
}
