% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{gs_item}
\alias{gs_item}
\title{GridStack Item}
\usage{
gs_item(
  ...,
  x = NULL,
  y = NULL,
  w = NULL,
  h = NULL,
  maxW = NULL,
  minW = NULL,
  maxH = NULL,
  minH = NULL,
  locked = NULL,
  noResize = NULL,
  noMove = NULL,
  id = NULL,
  class_item = NULL,
  class_content = NULL,
  style_item = NULL,
  style_content = NULL
)
}
\arguments{
\item{...}{Content of the item.}

\item{x, y}{(number) element position in column/row. Note: if one is missing this will autoPosition the item.}

\item{w, h}{(number) element size in column/row (default 1x1).}

\item{maxW, minW, maxH, minH}{element constraints in column/row (default none).}

\item{locked}{means another widget wouldn't be able to move it during dragging or resizing.
The widget can still be dragged or resized by the user. You need to add noResize and noMove attributes to completely lock the widget.}

\item{noResize}{disable element resizing.}

\item{noMove}{disable element moving.}

\item{id}{Good for quick identification (for example in change event).}

\item{class_item, class_content, style_item, style_content}{CSS class or CSS styles to apply to the item container or the content.}
}
\value{
A \code{list()} with a \code{shiny.tag} class.
}
\description{
GridStack Item
}
\examples{

library(gridstackr)

gridstack(
  gs_item("1", w = 4, h = 2),
  gs_item("Locked (can't be moved when other are moved)", w = 5, locked = TRUE),
  gs_item("No move", w = 5, x = 4, y = 2, noMove = TRUE),
  gs_item("No resize", h = 2, w = 3, noResize = TRUE),
  gs_item("Max width = 8", w = 6, x = 0, y = 3, maxW = 8),
  gs_item("Max height = 2", w = 6, x = 6, y = 3, maxH = 2)
)

}
